feat(testing): Expose Metrics interface rather (#3)
When consumers of lob/metrics are writing tests, they commonly
need to replace the implementation with a stub. Otherwise the
code under test will attempt to dial statsd, which will fail if
it's not running.

This commit updates metrics-go to expose an interface, rather
than a concrete implementation. This allows consumers to easily
drop in a stub during testing.